lz4: Avoid integer overflow in lz4_filter_read
Integer promotion stops at unsigned int, thus addition of 4 to a
uint32_t can lead to an integer overflow.

Use int64_t to support all values.

+++ libarchive/archive_read_support_filter_lz4.c
@@ -408,7 +408,7 @@ lz4_filter_read(struct archive_read_filter *self, const void **p)
 				    "Malformed lz4 data");
 				return (ARCHIVE_FATAL);
 			}
-			uint32_t skip_bytes = archive_le32dec(read_buf);
+			int64_t skip_bytes = archive_le32dec(read_buf);
 			__archive_read_filter_consume(self->upstream,
 				4 + skip_bytes);
 		} else {
